requirements.txt
fastapi
uvicorn[standard]
pydantic
pytesseract
pdf2image
pillow
opencv-python-headless
python-multipart
